Re-arrang code to separated files. Extract metadata and content from posts. Parse metadata into dictionary
