+++ rule-keeper/Dockerfile-rule-keeper
@@ -5,4 +5,4 @@ WORKDIR /usr/src/app
 COPY . .
 RUN pip install --no-cache-dir -r requirements.txt
 
-CMD [ "python", "./rule-keeper.py" ]
+CMD [ "python", "./rule_keeper.py" ]
