+++ rule-keeper/requirements.txt
@@ -1,1 +1,2 @@
 jellyfish==0.9.0
+PyYAML==6.0
